+++ libs/components/errors/src/app/docs/error/error-docs.component.ts
@@ -1,9 +1,0 @@
-import {
-  Component
-} from '@angular/core';
-
-@Component({
-  selector: 'app-error-docs',
-  templateUrl: './error-docs.component.html'
-})
-export class ErrorDocsComponent { }