+++ libs/components/errors/src/app/public/public_api.ts
@@ -1,3 +1,0 @@
-export * from './modules/error/error-modal-config';
-export * from './modules/error/error-modal.service';
-export * from './modules/error/error.module';